// front/src/providers/api-client.generated/model/appImageDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppImageDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    disabled?: boolean;
    name: string;
    size?: number;
    mimeType: string;
    physicalName?: string;
    width: number;
    height: number;
}